move TypeMatchError to it's own file
